+++ cmd/cron/add.go
@@ -7,6 +7,7 @@ import (
 	"strings"
 
 	"github.com/samcarswell/trochilus/config"
+	"github.com/samcarswell/trochilus/data"
 	"github.com/spf13/cobra"
 )
 
@@ -18,9 +19,16 @@ var addCmd = &cobra.Command{
 		if err != nil {
 			log.Fatalf("Could not get cron name %s", err)
 		}
+		notifyLog, err := cmd.Flags().GetBool("notify-log")
+		if err != nil {
+			log.Fatalf("Could not get notify-log %s", err)
+		}
 		queries := config.GetDatabase(cmd.Context())
 
-		newCronId, err := queries.CreateCron(cmd.Context(), cronName)
+		newCronId, err := queries.CreateCron(cmd.Context(), data.CreateCronParams{
+			Name:             cronName,
+			NotifyLogContent: notifyLog,
+		})
 		if err != nil {
 			if strings.Contains(err.Error(), "UNIQUE constraint failed: crons.name") {
 				log.Fatalf("Cron with name "+cronName+" already exists. %s", err)
@@ -34,7 +42,8 @@ var addCmd = &cobra.Command{
 
 func init() {
 	CronCmd.AddCommand(addCmd)
-	addCmd.Flags().String("name", "", "Cron Name")
+	addCmd.Flags().String("name", "", "Cron Name (required)")
+	addCmd.Flags().Bool("notify-log", false, "Includes log output in notification message (default false)")
 	if err := addCmd.MarkFlagRequired("name"); err != nil {
 		log.Fatalf("Unable to mark name as required %s", err)
 	}