+++ db/migrations/20251108020521_message_cron.sql
@@ -1,0 +1,7 @@
+-- migrate:up
+alter table crons
+add column notify_log_content boolean not null default false;
+
+-- migrate:down
+alter table crons
+drop column notify_log_content;